UI: Added more mouse speed levels in input config

+++ Core/Shared/KeyManager.h
@@ -30,7 +30,7 @@ public:
 	static void UpdateDevices();
 	
 	static void SetMouseMovement(int16_t x, int16_t y);
-	static MouseMovement GetMouseMovement(Emulator* emu, double mouseSensitivity);
+	static MouseMovement GetMouseMovement(Emulator* emu, uint32_t mouseSensitivity);
 	
 	static void SetMousePosition(Emulator* emu, double x, double y);
 	static MousePosition GetMousePosition();
